requirements.txt
google-cloud-vision>=3.7,<4
google-api-python-client>=2.100,<3
google-auth>=2.23,<3
google-genai>=1.0,<2
